Default Process Input Implementation - Made the processInput() implementation by PongGame the default for Game.
